pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "empire"
version = "1.0.0"
description = "Stochastic linear program for investments in the European power system."
authors = ["Stian Backe <stian.backe@ntnu.no>"]
maintainers = ["Stian Backe <stian.backe@ntnu.no>", "Martin Hjelmeland <martin.hjelmeland@ntnu.no>"]
readme = "README.md"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120

[tool.poetry.dependencies]
python = ">=3.7"
cloudpickle = ">=1.0.0"
openpyxl = ">=3.0.0"
pandas = ">=1.2"
pyomo = ">=6.0"
pyyaml = ">=5.1"
